Add pyspark versions of queries for spark-submit
The runSuite.pl script expects to be able to use spark-sql to submit
queries. The spark-submit command line utility is not present in the
Cloudera distribution because it does not support kerberized deployments.
By packaging queries as pyspark applications we can instead execute them
using the spark-submit command line utility.